requirements.txt
# requirements.txt

numpy
omegaconf
psutil
pyserial
pyyaml
